% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commas.R
\name{add_commas}
\alias{add_commas}
\title{print big numbers with commas}
\usage{
add_commas(x)
}
\arguments{
\item{x}{Numeric or character vector.}
}
\value{
String with commas in the right places.
}
\description{
print big numbers with commas
}
