% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_names.R
\name{add_names}
\alias{add_names}
\title{Adds names to data object}
\usage{
add_names(x, names)
}
\arguments{
\item{x}{Data}

\item{names}{Names.}
}
\value{
(Re)named data.
}
\description{
Adds names to data object wrapper of names<-
}
\examples{
add_names(mtcars, paste0("mtcars", seq_len(ncol(mtcars))))
}
