% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var.R
\name{add_var}
\alias{add_var}
\alias{add_vars}
\title{Mutates new variable, adding it to data frame}
\usage{
add_var(x, ...)

add_vars(x, ...)
}
\arguments{
\item{x}{Data frame to be added to.}

\item{...}{Named variables to add to data frame}
}
\value{
A data frame.
}
\description{
Adds new variable column to data frame

Adds additional columns for every named variable provided
}
\examples{

## add to vars to mtcars data set
add_vars(mtcars[, 1:4], a = TRUE, b = rnorm(32))

}
