% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{any_recursive}
\alias{any_recursive}
\title{any_recursive}
\usage{
any_recursive(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
Logical
}
\description{
Tests whether any element is recursive.
}
\examples{
## list containing data frame
lst <- list(
  a = c(1, 2, 3),
  b = "asdf",
  c = mtcars
)
any_recursive(lst)

## list with non-recursive columns
lst <- list(
  a = c(1, 2, 3),
  b = "asdf",
  c = c(TRUE, FALSE, TRUE, FALSE)
)
any_recursive(lst)

## data frame containing data frame
dat <- data.frame(
  a = c(1, 2, 3),
  b = c("a", "b", "c"),
  c = I(list(mtcars, NA, "blah"))
)
any_recursive(dat)

## data frame with non-recursive columns
dat <- data.frame(
  a = c(1, 2, 3),
  b = c("a", "b", "c"),
  c = c(TRUE, FALSE, TRUE)
)
any_recursive(dat)
}
