% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_citation.R
\name{apa_citation}
\alias{apa_citation}
\title{APA citation of R package}
\usage{
apa_citation(pkg)
}
\arguments{
\item{pkg}{Name of package (quoted string)}
}
\value{
A character vector of the APA citation. If on Mac or PC, then also
  it's stored to the system's clipboard.
}
\description{
Returns an APA-formatting citation of an R package
}
\examples{

## cite this pacakge
apa_citation("tfse")

}
