% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_call.R
\name{api_call}
\alias{api_call}
\title{Build url for HTTP request}
\usage{
api_call(url, ...)
}
\arguments{
\item{url}{Base URL should include base, path, version, etc.}

\item{...}{Named arguments beyond url are treated as named
query parameters to be passed along in the API call.}
}
\value{
Character string with API request URL.
}
\description{
Builds HTTP request for a given API call.
}
\examples{

## request URL
rurl <- api_call(
  "https://api.twitter.com/1.1/search/tweets.json",
  q = "rstats",
  count = 100
)

\dontrun{
## send http request (this should return error without access token)
r <- httr::GET(rurl)

## view returned content
httr::content(r)
}

}
