% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_nth.R
\name{append_nth}
\alias{append_nth}
\title{Append integers with appopriate suffix}
\usage{
append_nth(x)
}
\arguments{
\item{x}{Integer or integerish vector.}
}
\value{
A character vector with integers converted to "th"
}
\description{
Adds "st", "nd", "rd", or "th" suffix to integers
}
\examples{

## randomly sample ten values from 1:100 and add suffix
append_nth(sample(1:100, 10))

}
