% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_lines.R
\name{break_lines}
\alias{break_lines}
\title{break_lines}
\usage{
break_lines(x, n = 80, sep = "\\\\. ", collapse = "\\n")
}
\arguments{
\item{x}{Input text}

\item{n}{Number of characters (width) on which to truncate given text}

\item{sep}{Separator used for subunits.}

\item{collapse}{When text is recombined into single string, this value is used
as the sepator. Defaults to three line breaks.}
}
\value{
Output with line breaks optimized for n width.
}
\description{
Truncates lines to a supplied width value
}
