% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_gh.R
\name{browse_gh}
\alias{browse_gh}
\title{Open Github repo}
\usage{
browse_gh(repo = NULL)
}
\arguments{
\item{repo}{Name of github repo. If null (the default), function opens to
github.com. For several popular repos, simply supply the name of the pkg/
repo and the user/org will be imputed. Otherwise, supply the user/repo.}
}
\value{
Opens browser to appropriate Github.com page
}
\description{
A github.com-specific wrapper around browseURL
}
