% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{c.string}
\alias{c.string}
\title{c.string}
\usage{
\method{c}{string}(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Printed output via cat() function for copy/paste of long strings
}
\description{
c.string
}

