% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numint.R
\name{can_int}
\alias{can_int}
\title{Determines whether vector can be converted to integer}
\usage{
can_int(x)
}
\arguments{
\item{x}{Vector to test whether it can be converted to integer.}
}
\value{
Logical value of true or false.
}
\description{
Converts vector (if possible) into integer and determines if it makes
sense to do so.
}
