% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numint.R
\name{can_num}
\alias{can_num}
\title{Determines whether vector can be converted to numeric}
\usage{
can_num(x)
}
\arguments{
\item{x}{Vector to test whether it can be converted to numeric.}
}
\value{
Logical value of true or false.
}
\description{
Converts vector (if possible) into numeric and determines if it makes
sense to do so.
}
