% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_labels.R
\name{capitalize_labels}
\alias{capitalize_labels}
\title{Capitalize labels}
\usage{
capitalize_labels(x)
}
\arguments{
\item{x}{Input object, character, data frame, list, ggplot, etc.}
}
\value{
Returns the same class as input
}
\description{
Convert text into title case (first word only)
}
