% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.R
\name{carbon.sh}
\alias{carbon.sh}
\title{Create carbon.sh image in browser}
\usage{
carbon.sh(x)
}
\arguments{
\item{x}{Code to export to carbon for image capture}
}
\value{
Opens browser to carbon.sh with supplied code
}
\description{
Create carbon.sh image in browser
}
