% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_line.R
\name{cat_lines}
\alias{cat_lines}
\alias{cat_line}
\title{Combine and print text}
\usage{
cat_lines(...)

cat_line(...)
}
\arguments{
\item{...}{Text to print and/or named arguments such as sep, collapse, fill,
indent, postdent.}
}
\value{
Prints text and returns NULL similar to the base cat function
}
\description{
\code{cat_lines}: concatenate/print \strong{multiple} line

\code{cat_line}: concatenate/print \strong{single} line
}
