% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{col2hex}
\alias{col2hex}
\title{col2hex}
\usage{
col2hex(color, alpha = NULL)
}
\arguments{
\item{color}{One or more named colors. Also accepts hex colors (useful for
adding alpha values).}

\item{alpha}{Alpha level ranging form 1.0 (opaque) to 0.0 (transparent). The
default (NULL) omits alpha columns from the returned hexadecimal
characters.}
}
\value{
Hexadecimal color codes
}
\description{
Converts named colors to hex (html) codes.
}
