% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_function.R
\name{copy_function}
\alias{copy_function}
\title{Copy function to clipboard}
\usage{
copy_function(...)
}
\arguments{
\item{...}{Functions to write to clipboard. It's best to name these.}
}
\value{
Writes to clipboard (ready to paste)
}
\description{
Copy the function code to clipboard (ready to paste)
}
