% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.lookup.R
\name{data_frame_lookup}
\alias{data_frame_lookup}
\title{data_frame_lookup}
\usage{
data_frame_lookup(x)
}
\arguments{
\item{x}{json resposne object from user lookup Twitter API call.}
}
\value{
data frame
}
\description{
data_frame_lookup
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more information on using Twitter's API.
}

