% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\name{data_set}
\alias{data_set}
\title{Create data set}
\usage{
data_set(...)
}
\arguments{
\item{...}{Data to be converted to data frame}
}
\value{
A tibble data frame
}
\description{
Simple way to make data frames
}
