% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_read_csv}
\alias{db_read_csv}
\title{read CSV from dropbox}
\usage{
db_read_csv(file, ...)
}
\arguments{
\item{file}{Path/file name in dropbox folder to read.}

\item{...}{Other args passed to readr::read_csv}
}
\value{
Reads CSV file from dropbox.
}
\description{
read CSV from dropbox
}
