% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_readRDS}
\alias{db_readRDS}
\title{read rds from dropbox}
\usage{
db_readRDS(file)
}
\arguments{
\item{file}{Path/file name in dropbox folder to read.}
}
\value{
Reads rds file from dropbox.
}
\description{
read rds from dropbox
}
