% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_saveRDS}
\alias{db_saveRDS}
\title{write rds to dropbox}
\usage{
db_saveRDS(x, file)
}
\arguments{
\item{x}{Data object to save}

\item{file}{Path/file name in dropbox folder to save data.}
}
\value{
Saves x as rds file in specified dropbox folder
}
\description{
write rds to dropbox
}
