% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_write_csv}
\alias{db_write_csv}
\title{write csv to dropbox}
\usage{
db_write_csv(x, file)
}
\arguments{
\item{x}{Data object to save}

\item{file}{Path/file name in dropbox folder to save data.}
}
\value{
Saves x as CSV file in specified dropbox folder
}
\description{
write csv to dropbox
}
