% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_get_var}
\alias{desc_get_var}
\alias{desc_get_url}
\alias{desc_gh_repo}
\title{Description variables}
\usage{
desc_get_var(pkg, field = NULL)

desc_get_url(pkg)

desc_gh_repo(pkg)
}
\arguments{
\item{pkg}{Character string, name of package}

\item{field}{Character string, name of field/key}
}
\value{
Data frame
}
\description{
Get values from package DESCRIPTION
}
