% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{.Renviron}
\alias{.Renviron}
\title{.Renviron file}
\usage{
.Renviron()
}
\value{
Returns path to .Renviron file
}
\description{
Gets path to .Renviron file
}
\details{
Checks local directory first and then checks home directory
}
