% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_names.R
\name{edit_names}
\alias{edit_names}
\title{edit names}
\usage{
edit_names(x, expr)
}
\arguments{
\item{x}{Object with names to be edited}

\item{expr}{Expression to perform on names. IMPORTANT: assumes names are .x}
}
\value{
Returns renamed x object
}
\description{
Edit names, reassign them, and then return object
}
