% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emj.R
\docType{data}
\name{emj}
\alias{emj}
\title{Emojis for creating print messages}
\format{A data frame with 1 row and 30 variables:}
\usage{
emj
}
\description{
A dataset containing emojis useful for message printing.
}
\keyword{datasets}
