% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{enc2ascii}
\alias{enc2ascii}
\title{emc2ascii}
\usage{
enc2ascii(x, y = "")
}
\arguments{
\item{x}{Text, a character vector.}

\item{y}{Replacement for non-ascii characters. Defaults to "" (blank).}
}
\value{
ascii friendly text
}
\description{
Makes text to ascii friendly.
}
