% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_function.R
\name{file_edit}
\alias{file_edit}
\title{Open file in text editor}
\usage{
file_edit(file)
}
\arguments{
\item{file}{Name of file(s) to open.}
}
\value{
Opens file and returns invisible file name.
}
\description{
Opens file in current text editor
}
\details{
Should open using system default or current text editor.
}
