% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_names.R
\name{fully_qualified_namespaces}
\alias{fully_qualified_namespaces}
\title{find namespaces from fully qualified references in a directory}
\usage{
fully_qualified_namespaces(dir = ".")
}
\arguments{
\item{dir}{Name of directory to search, defaults to current dir.}
}
\value{
Unique list of namespaces
}
\description{
find namespaces from fully qualified references in a directory
}
