% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{gather}
\alias{gather}
\title{Convert data from wide to long}
\usage{
gather(data, key = "key", value = "value", ..., na.rm = FALSE,
  convert = FALSE, factor_key = FALSE)
}
\description{
Convert data from wide to long
}
\keyword{internal}
