% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_api}
\alias{get_api}
\title{get_api}
\usage{
get_api(url, token = NULL)
}
\arguments{
\item{url}{API url address.}
}
\value{
Response formatted as nested list.
Assumes response object is json object.
}
\description{
get_api
}
\details{
httr jsonlite
}

