% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.followers.R
\name{get_followers}
\alias{get_followers}
\title{get_followers}
\usage{
get_followers(user, token, page = "-1")
}
\arguments{
\item{token}{OAuth token (1.0 or 2.0)}

\item{page}{Default \code{page = -1} specifies first page of json results. Other pages specified via cursor values supplied by Twitter API response object.}

\item{screen_name}{Screen name of target user.}
}
\value{
user ids
}
\description{
get_followers
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more information on using Twitter's API.
}

