% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.followers.R
\name{get_followers_max}
\alias{get_followers_max}
\title{get_followers_max}
\usage{
get_followers_max(user, tokens)
}
\arguments{
\item{user}{Screen name or user id of target user}

\item{tokens}{OAuth tokens (1.0 or 2.0)}
}
\value{
user ids
}
\description{
get_followers_max
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more
  information on using Twitter's API.
}

