% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.followers.R
\name{get_followerslist}
\alias{get_followerslist}
\title{get_followerslist}
\usage{
get_followerslist(user, token, page = "-1")
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{token}{OAuth token (1.0 or 2.0)}

\item{page}{Default \code{page = -1} specifies first page of json results.
Other pages specified via cursor values supplied by Twitter API response
object.}
}
\value{
json user object (nested list)
}
\description{
Returns a cursored collection of user objects for users following the
specified user.
}
\seealso{
See \url{https://api.twitter.com/1.1/followers/list.json}.
}

