% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.friends.R
\name{get_friends_max}
\alias{get_friends_max}
\title{get_friends_max}
\usage{
get_friends_max(user_ids, tokens, n = 1, stringify = TRUE, verbose = TRUE)
}
\arguments{
\item{tokens}{OAuth tokens (1.0 or 2.0)}

\item{stringify}{logical, indicating whether to return user ids as strings
(some ids are too long to be read as numeric). Defaults to \code{TRUE}.}

\item{ids}{Data frame with column name "screen_name"}

\item{start}{Starting value (nth user id)}
}
\value{
friends List of user ids each user follows.
}
\description{
get_friends_max
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more
  information on using Twitter's API.
}

