% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.friends.R
\name{get_friends_ply}
\alias{get_friends_ply}
\title{get_friends_ply}
\usage{
get_friends_ply(user_ids, tokens, start = 1)
}
\arguments{
\item{user_ids}{Data frame with column name "screen_name"}

\item{tokens}{OAuth tokens (1.0 or 2.0)}

\item{start}{Starting value (nth user id)}
}
\value{
friends List of user ids each user follows.
}
\description{
lapply() version of get_friends_max()
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more
  information on using Twitter's API.
}

