% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.functions.R
\name{get_friends_timepoint}
\alias{get_friends_timepoint}
\title{get_friends_timepoint}
\usage{
get_friends_timepoint(followers, tokens, screen_name, N = 600)
}
\arguments{
\item{followers}{Data frame with user ids in column one}

\item{tokens}{OAuth tokens (1.0 or 2.0)}
}
\value{
response object
}
\description{
get_friends_timepoint
}

