% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.friends.R
\name{get_friendships}
\alias{get_friendships}
\title{get_friendships}
\usage{
get_friendships(source, target, token)
}
\arguments{
\item{source}{Screen name or user id of source user.}

\item{target}{Screen name or user id of target user.}

\item{token}{OAuth token (1.0 or 2.0)}
}
\value{
json object (nested list)
}
\description{
Returns detailed information about the relationship between two arbitrary
users.
}
\seealso{
See \url{https://api.twitter.com/1.1/friendships/show.json}.
}

