% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.friends.R
\name{get_friendslist}
\alias{get_friendslist}
\title{get_friendslist}
\usage{
get_friendslist(user, token, page = "-1")
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{token}{OAuth token (1.0 or 2.0)}

\item{page}{Default \code{page = -1} specifies first page of json results.
Other pages specified via cursor values supplied by Twitter API response
object.}
}
\value{
json user object (nested list)
}
\description{
Returns a cursored collection of user objects for every user the specified
user is following (otherwise known as their “friends”).
}
\seealso{
See \url{https://api.twitter.com/1.1/friends/list.json}.
}

