% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.list.R
\name{get_list}
\alias{get_list}
\title{get_list}
\usage{
get_list(slug, owner, token, count = 5000, cursor = "-1")
}
\arguments{
\item{slug}{Name of Twitter list}

\item{owner}{Screen name of list owner}

\item{token}{An OAuth token (1.0 or 2.0)}

\item{count}{Maximum number of users to return (cannot be higher than 5000).}

\item{cursor}{Select next or previous page or results using cursor value from returned json object}
}
\value{
json response object as nested list
}
\description{
get_list
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more information on using Twitter's API.
}

