% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.functions.R
\name{get_lookup}
\alias{get_lookup}
\title{get_lookup}
\usage{
get_lookup(users, token)
}
\arguments{
\item{users}{User ids of target user.}

\item{tokens}{OAuth tokens (1.0 or 2.0)
See \url{https://dev.twitter.com/overview/documentation} for more information on using Twitter's API.}
}
\value{
response object
}
\description{
get_lookup
}

