% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.functions.R
\name{get_lookup_max}
\alias{get_lookup_max}
\title{get_lookup_max}
\usage{
get_lookup_max(ids, tokens)
}
\arguments{
\item{ids}{User ids of target user.}

\item{tokens}{OAuth tokens (1.0 or 2.0)}
}
\value{
response object
}
\description{
get_lookup_max
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more information on using Twitter's API.
}

