% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.package.pdf.R
\name{get_package_pdf}
\alias{get_package_pdf}
\title{get_package_pdf}
\usage{
get_package_pdf(package, update = FALSE, open = TRUE)
}
\arguments{
\item{package}{Name of target package.}

\item{update}{logical, indicating whether to replace old manual}

\item{open}{logical, indicating whether to automatically view preview of pdf}
}
\description{
get_package_pdf
}

