% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.functions.R
\name{get_token}
\alias{get_token}
\title{get_token}
\usage{
get_token(app, consumer_key, consumer_secret)
}
\arguments{
\item{app}{Name of user created Twitter application}

\item{consumer_key}{Application API key}

\item{consumer_secret}{Application API secret User-owned app must have
\code{Read and write} access level and \code{Callback URL} of
\code{http://127.0.0.1:1410}.}
}
\value{
twitter oauth 1.0 token
}
\description{
get_token
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more
  information on using Twitter's API.
}

