% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_var}
\alias{get_var}
\title{get_var}
\usage{
get_var(x, ...)
}
\arguments{
\item{x}{Data object. Particularly useful for recursive lists.}

\item{...}{Names at each successive level from which to dive and return.}
}
\value{
Returns the last named element found in the data object.
}
\description{
Looks for and returns element of given name
}
\examples{
## generate nested list
lst <- list(
  I = 1,
  II = list(
    A = 1,
    B = list(
      a = 1,
      b = list(
        v1 = 1,
        v2 = 2
      )
    )
  )
)
lst <- replicate(2, list(lst))

## fetch and return all v2's
get_var(lst, "II", "B", "v2")
}
