% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.wave.R
\name{get_wave}
\alias{get_wave}
\title{get_wave}
\usage{
get_wave(ids, tokens)
}
\arguments{
\item{ids}{Vector of user ids}
}
\value{
friends List of user ids each user follows.
}
\description{
get_wave
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more information on using Twitter's API.
}

