% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.wave.R
\name{get_wave_data}
\alias{get_wave_data}
\title{get_wave_data}
\usage{
get_wave_data(wave)
}
\arguments{
\item{wave}{Wave number}
}
\value{
data Saved wave data and updated running data set
}
\description{
get_wave_data
}

