% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghinstall.R
\name{ghinstall}
\alias{ghinstall}
\title{Install package from Github}
\usage{
ghinstall(repo)
}
\arguments{
\item{repo}{Name of repo e.g., 'mkearney/warcraft'}
}
\value{
Installs package to default library
}
\description{
Installs an R package from Github
}
