% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_raw.R
\name{github_raw}
\alias{github_raw}
\title{Generate link to raw Github file}
\usage{
github_raw(file, repo = NULL)
}
\arguments{
\item{file}{Name of desired file; if \code{repo} is NULL, then this value
should also provide repo information, i.e., owner and name of the
repository–e.g., \code{"owner/repo/file.ext"}. Alternatively, user may
supply the URL to the file as it appears in a web browser, e.g.,
\code{"https://github.com/mkearney/driven-snow/blob/master/theme/driven-snow.rstheme"}}

\item{repo}{Repository name information formatted as username/repo. If this
information is provided in the value supplied to \code{file} then leave
this as NULL (the default)}
}
\value{
Returns the URL path to the raw version of the file.
}
\description{
Converts Github path/repo/file information into a link to the raw version of
the file
}
