% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{\%nin\%}
\alias{\%nin\%}
\title{not in}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{Left hand side}

\item{rhs}{Right hand side}
}
\value{
Logical indicating whether lhs are NOT in rhs. For more
  info on the output vector.
}
\description{
A more convenient way to say no lhs in rhs
}
\details{
Number of returned logical values is equal to the length
  or number of columns of LHS (behaves the same as base in
  function).
}
\examples{
## is "a" in the alphabet?
"a" \%nin\% letters

## vector of "elephant" letters?
elephant <- strsplit("elephant", "")[[1]]

## letters a b and c
abc <- c("a", "b", "c")

## are  "a", "b", or "c" not in
abc \%nin\% elephant

}
