% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paste.R
\name{\%P\%}
\alias{\%P\%}
\title{Paste grapes}
\usage{
lhs \%P\% rhs
}
\arguments{
\item{lhs}{Left hand side presumably character string}

\item{rhs}{Right hand side presumably another character string}
}
\value{
A pasted together (with no space) string(s)
}
\description{
Paste0 strings together with grapes (inverse/inside out function call)
}
