% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmatches_.R
\name{gregexpr_}
\alias{gregexpr_}
\alias{regexpr_}
\title{smart gregexpr wrapper}
\usage{
gregexpr_(x, pat, ...)

regexpr_(x, pat, ...)
}
\arguments{
\item{x}{Input text}

\item{pat}{Reg ex pattern}

\item{...}{Other args passed to base (g)regexpr}
}
\value{
Pattern match positions
}
\description{
smart gregexpr wrapper
}
