% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{has_factor_potential}
\alias{has_factor_potential}
\title{has_factor_potential}
\usage{
has_factor_potential(x, p_uq = 0.9, max_chars = 60)
}
\arguments{
\item{x}{Vector}

\item{p_uq}{Proportion of cases that represent unique values. If this value
is less than one, than ID vars will be returned FALSE, and so on.}

\item{max_chars}{Number of chars from which the maximum acceptable
MEAN string size should be allowed.}
}
\value{
Logical.
}
\description{
Tests whether vector has factor potential
}
