% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{has_name_}
\alias{has_name_}
\title{has_name_}
\usage{
has_name_(x, ...)
}
\arguments{
\item{x}{Data object. This function is not recursive, so must be applied to
object at intended level of inspection.}

\item{...}{Names to look for in data object.}
}
\value{
Logical reflecting whether data contains provided name(s)
}
\description{
Tests whether object contains name(s)
}
\details{
If multiple names are provided, the function will only return TRUE if
  all names are present.
}
\examples{
## data set
dat <- data.frame(a = 1, b = 2, c = 3)

## does dat contain named element "a"
has_name_(dat, "a")

## what about "q"?
has_name_(dat, "q")

## what about both "a" and "b"?
has_name_(dat, "a", "b")

#' ## what about "a", "b", and "q"?
has_name_(dat, "a", "b", "q")
}
