% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.R
\name{hd}
\alias{hd}
\title{Head data}
\usage{
hd(x, i = 10)
}
\arguments{
\item{x}{Input}

\item{i}{Number of elements to return from first, defaults to 10.}
}
\value{
Data with first i elements.
}
\description{
Returns first i elements of data object.
}
