% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexdf.R
\name{hexdf}
\alias{hexdf}
\title{Generate a hexagon shaped data frame}
\usage{
hexdf(size = 1, mid = c(0, 0), angle = 90)
}
\arguments{
\item{size}{The size of hexagon (one half of the width).}

\item{mid}{The xy midpoint, defaulting to c(0, 0)}

\item{angle}{The angle at which to rotate the hexagon. The default of 90 results in the top and bottoms of the hexagon coming to a point}
}
\value{
A tibble data frame
}
\description{
Returns xy coordinates for hexagon shape using hex sticker dimensions
}
\examples{
## plot hexagon using function defaults
hex1 <- hexdf()
with(hex1, plot(x, y, type = "l"))

## create larger hexagon to the lower right
hex2 <- hexdf(2, mid = c(2.5, -2))

## plot both hex objects
ggplot2::ggplot(hex1, ggplot2::aes(x, y)) +
  ggplot2::geom_polygon(fill = "blue") +
  ggplot2::geom_polygon(data = hex2, fill = "red")

## create hex pkg sticker

## some points (dots)
pts <- data.frame(
  x = runif(30, -.5, .5),
  y = runif(30, -.05, .5))

## use ggplot to create sticker
ggplot2::ggplot(hex1,ggplot2::aes(x, y)) +
  ggplot2::geom_polygon(fill = "#2244aa", colour = "#001050", size = 4) +
  ggplot2::geom_point(data = pts, shape = 21, size = 10,
    colour = "#001050", fill = "white") +
  ggplot2::theme_void() +
  ggplot2::annotate("text", 0, -.4, label = "dots", colour = "white",  size = 42,
    fontface = "bold")
## pass to ggplot
}
