% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{home}
\alias{home}
\title{Home directory}
\usage{
home()
}
\value{
Returns system/user's default home directory
}
\description{
Gets user's home directory
}
\details{
Looks for "HOME" environment variable and/or normalizes the tilde
  path
}
