% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pkg.R
\name{install_pkg}
\alias{install_pkg}
\title{Install package}
\usage{
install_pkg(pkg = ".", ...)
}
\arguments{
\item{pkg}{Name/path to package defaults to current directory}

\item{...}{Passed to package build}
}
\value{
Installs locally
}
\description{
A wrapper function to simplify devtools install
}
