% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.functions.R
\name{is_screen_name}
\alias{is_screen_name}
\title{is_screen_name}
\usage{
is_screen_name(x)
}
\arguments{
\item{x}{Twitter user id or screen name}
}
\value{
logical value indicating whether object is screen name [or user ID]
}
\description{
is_screen_name
}

