% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.R
\name{sem_fit}
\alias{sem_fit}
\alias{par_table}
\title{SEM fit statistics}
\usage{
sem_fit(x, digits = 3, scaled = TRUE)

par_table(x)
}
\arguments{
\item{x}{Object of class lavaan.}

\item{digits}{Number of digits to round to, Defaults to 3.}

\item{scaled}{Logical indicating whether to return scaled fit indices if
available.}

\item{x}{Object of class lavaan}
}
\value{
Fit statistics tibble.

Tibble of parameter estimates
}
\description{
Returns data frame of SEM fit statistics

SEM parameter table
}
\examples{
## specify model
m1 <- 'e =~ cyl + hp + wt\\nmpg ~ e + gear'

## function to normalize data
normalize <- function(x) (x - mean(x, na.rm = TRUE)) / sd(x, na.rm = TRUE)

## fit model
f1 <- lavaan::sem(
  m1,
  data = tibble::as_tibble(dplyr::mutate_if(mtcars, is.numeric, normalize)),
  std.lv = TRUE,
  estimator = "mlr"
)

## view fit stats
sem_fit(f1)

## view par estimates
par_table(f1)

}
