% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{lib}
\alias{lib}
\title{Silently load R packages}
\usage{
lib(x)
}
\arguments{
\item{x}{Name of library. Accepts unquoted or quoted text.}
}
\value{
Loads package namespace.
}
\description{
Shortcut for library that is completely silent
}
\examples{
## load the tidyverse without all the messages
lib(tidyverse)
}
