% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location2state.R
\name{location2state}
\alias{location2state}
\title{Convert Twitter locations to U.S. state names}
\usage{
location2state(x)
}
\arguments{
\item{x}{Character vector of Twitter user location data.}
}
\value{
Character vector of same length with U.S. names (or a
  generic "United States") for values based on best guess for each
  user. Cases with insufficient information are returned as NAs.
}
\description{
Estimates U.S. state given Twitter user location data.
}
