% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{make_citation}
\alias{make_citation}
\title{make_citation}
\usage{
make_citation(pkg = ".")
}
\arguments{
\item{pkg}{Name of package. Defaults to current file name.}
}
\value{
Saves updated CITATION file.
}
\description{
Makes (or updates) CITATION file.
}
