% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{make_package}
\alias{make_package}
\title{make_package}
\usage{
make_package(update = NULL, pkg = ".", load_all = TRUE)
}
\arguments{
\item{update}{Optional, indicating the size of change so as to automate
updating of the package version number.}

\item{pkg}{Name of package. Defaults to current file name.}

\item{load_all}{Logical indicating whether or not to load all functions
on exit.}
}
\description{
Document and load R package (devtools wrapper)
}
