% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-purrr.R
\name{map}
\alias{map}
\alias{map_lgl}
\alias{map_int}
\alias{map_dbl}
\alias{map_chr}
\alias{map_cpl}
\title{Base version of map over function}
\usage{
map(.f, .x, ...)

map_lgl(.f, .x, ...)

map_int(.f, .x, ...)

map_dbl(.f, .x, ...)

map_chr(.f, .x, ...)

map_cpl(.f, .x, ...)
}
\arguments{
\item{.f}{function}

\item{.x}{first param value}

\item{...}{additional args}
}
\value{
Data returned as list vector
}
\description{
Base version of map over function equivalent of lapply
}
\examples{

map(is.numeric, mtcars)
}
