% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_}
\alias{map_}
\alias{map_chr_}
\alias{map_dbl_}
\alias{map_int_}
\alias{map_lgl_}
\title{map_}
\usage{
map_(f, ...)

map_chr_(f, ...)

map_dbl_(f, ...)

map_int_(f, ...)

map_lgl_(f, ...)
}
\arguments{
\item{f}{Function}

\item{...}{Other args passed to function, i.e., named parameters
specific to f function, and also the \code{MoreArgs} argument
from \code{mapply}.}

\item{f}{Function}

\item{...}{Other args passed to function, i.e., named parameters
specific to f function, and also the \code{MoreArgs} argument
from \code{mapply}.}

\item{f}{Function}

\item{...}{Other args passed to function, i.e., named parameters
specific to f function, and also the \code{MoreArgs} argument
from \code{mapply}.}

\item{f}{Function}

\item{...}{Other args passed to function, i.e., named parameters
specific to f function, and also the \code{MoreArgs} argument
from \code{mapply}.}

\item{f}{Function}

\item{...}{Other args passed to function, i.e., named parameters
specific to f function, and also the \code{MoreArgs} argument
from \code{mapply}.}
}
\value{
List

Vector type character

Vector type double

Vector type integer

Vector type logical.
}
\description{
maps function and returns list

maps function and returns character vector

maps function and returns double vector

maps function and returns integer vector

maps function and returns logical vector
}
\details{
Yes, this is identical to base \code{Map} but it also
  hardwires USE.NAMES to FALSE.
}
\examples{
map_(as.data.frame, mtcars)
x <- c("A", "brown", "car", "drove", "eighty", "four", "gallons", "how")
map_chr_(function(x) substr(x, 1, 1), x)
x <- rnorm(10)
map_dbl_(round, x, 2)
x <- rnorm(10, 100, 5)
map_int_(as.integer, x, 2)
map_lgl_(is.numeric, list(rnorm(10), rnorm(10)))
}
