% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_chr}
\alias{map_chr}
\title{map_chr}
\usage{
map_chr(f, ...)
}
\arguments{
\item{f}{Function}

\item{...}{Other args passed to function.}
}
\value{
Vector type character
}
\description{
map_chr
}
\examples{
x <- c("A", "brown", "car", "drove", "eighty", "four", "gallons", "how")
map_chr(function(x) substr(x, 1, 1), x)
}
