% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_dbl}
\alias{map_dbl}
\title{map_dbl}
\usage{
map_dbl(f, ...)
}
\arguments{
\item{f}{Function}

\item{...}{Other args passed to function.}
}
\value{
Vector type double
}
\description{
map_dbl
}
\examples{
x <- rnorm(10)
map_dbl(round, x, 2)
}
