% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{max_na}
\alias{max_na}
\title{Filter based on proportion of missing data}
\usage{
max_na(x, max = 0.05)
}
\arguments{
\item{x}{Input data frame or matrix.}

\item{max}{Maximum proportion of missingnesses allowed. Columns with higher
proportions of missingness compared to this value will be dropped. Columns
will only be returned if they have 1 - max proportion non-missing. This
value must be between 0-1. It defaults to .05.}
}
\value{
Data frame or matrix with columns with less than or equal to the
  max allowed proportion of missingness.
}
\description{
Returns columns that have less than or equal to a specified amount of missingness.
}
