% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{menuline}
\alias{menuline}
\title{menuline}
\usage{
menuline(q, a)
}
\arguments{
\item{q}{Question to be asked in interactive session.}

\item{a}{Answer choices.}
}
\value{
Selection provided by user in interactive session.
}
\description{
Creates interactive multiple choice question.
}
