% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{MIT_license}
\alias{MIT_license}
\title{MIT_license}
\usage{
MIT_license(pkg = ".")
}
\arguments{
\item{pkg}{Name of package (basename of wd)}
}
\value{
Saves license file.
}
\description{
Creates MIT license
}
