% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mmap_lgl}
\alias{mmap_lgl}
\title{map_lgl}
\usage{
mmap_lgl(f, ...)
}
\arguments{
\item{f}{Function}

\item{...}{Other args passed to function.}
}
\value{
Vector type logical.
}
\description{
map_lgl
}
\examples{
map_lgl(is.numeric, list(rnorm(10), rnorm(10)))
}
