% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{n_uq}
\alias{n_uq}
\title{n_uq}
\usage{
n_uq(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Integer number of distinct elements.
}
\description{
Number of distinct elements.
}
\examples{
n_uq(sample(1:10, 5, replace = TRUE))
}
