% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in.R
\name{nin}
\alias{nin}
\alias{yin}
\title{Return lhs values not in rhs values}
\usage{
nin(lhs, rhs, value = TRUE)

yin(lhs, rhs, value = TRUE)
}
\arguments{
\item{lhs}{Values to check whether they are/not contained in the other}

\item{rhs}{Values to use as the reference}

\item{value}{Logical indicating whether to return the value or a logical vector}
}
\description{
Return lhs values not in rhs values
}
\examples{

## a, b, zz in alphabet letters
yin(c("a", "b", "zz"), letters)

## a, b, zz NOT in alphabet letters
nin(c("a", "b", "zz"), letters)

}
