% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{parse_json}
\alias{parse_json}
\title{grab json}
\usage{
parse_json(x)
}
\arguments{
\item{x}{input}
}
\value{
if json found then parsed list
}
\description{
grab json
}
\examples{
\dontrun{
## scrape espn
e <- parse_json("http:/www.espn.com/")
str(e, 3)

## convert html text
parse_json('<html>{\\"mpg\\":21,\\"cyl\\":6,\\"disp\\":160,\\"lgl\\":false}</html>')
}
}
