% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.objects.R
\name{parse_status_entities}
\alias{parse_status_entities}
\title{parse_status_entities}
\usage{
parse_status_entities(x)
}
\arguments{
\item{x}{statuses entities object}
}
\description{
parse_status_entities
}

