% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_time.R
\name{parse_to_secs}
\alias{parse_to_secs}
\title{parse_to_secs}
\usage{
parse_to_secs(x)
}
\arguments{
\item{x}{Unit of time in secs, mins, hours, days, weeks, etc.}
}
\value{
Number of seconds
}
\description{
Verbs object
}
