% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_function.R
\name{paste_collapse}
\alias{paste_collapse}
\title{Paste collapse}
\usage{
paste_collapse(...)
}
\arguments{
\item{...}{One or more character strings to paste together with paste0 and
collapse equal to ""}
}
\value{
A single string collapsed and separated with empty spaces
}
\description{
Paste with sep and collapse set to empty.
}
