% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{pbcopy}
\alias{pbcopy}
\title{pbcopy}
\usage{
pbcopy(x)
}
\arguments{
\item{x}{Input passed to cat function.}
}
\value{
Prints x to clipboard.
}
\description{
Adds input to clipboard for pasting
}
\examples{
## alphabet as string
pbcopy(paste(letters, collapse = ""))
## paste e.g., C-v
}
