% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{pfuns}
\alias{pfuns}
\title{funs}
\usage{
pfuns(names)
}
\arguments{
\item{names}{Names of functions.}
}
\value{
Skeleton function code copied to clipboard.
}
\description{
Pastes fun skeletons into clipboard
}
