% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{pkg_website}
\alias{pkg_website}
\title{Updates pkg website using pkgdown}
\usage{
pkg_website()
}
\description{
Updates pkg website using pkgdown
}
\details{
If custom css is supplied (must be saved as custom.css in the
  docs folder), then it will be used to supplement the default pkgdown
  theme.
}
