% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-purrr.R
\name{pluck}
\alias{pluck}
\alias{pluck_lgl}
\alias{pluck_int}
\alias{pluck_dbl}
\alias{pluck_chr}
\alias{pluck_cpl}
\title{Pluck elements from list}
\usage{
pluck(.x, .f)

pluck_lgl(.x, .f)

pluck_int(.x, .f)

pluck_dbl(.x, .f)

pluck_chr(.x, .f)

pluck_cpl(.x, .f)
}
\arguments{
\item{.x}{Data}

\item{.f}{Named data object or integer position number}
}
\value{
Data plucked from input
}
\description{
Extracts from list named or integer position data.
}
\examples{
pluck(list(mtcars), "cyl")

}
