% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_paste.R
\name{+.character}
\alias{+.character}
\title{Add (paste) together character vectors}
\usage{
\method{+}{character}(e1, e2)
}
\arguments{
\item{e1}{lhs}

\item{e2}{rhs}
}
\value{
Computed value
}
\description{
Paste or paste0 character vectors using the plus operator.
}
\examples{
## normal plus operator
20 + 1
Sys.Date() + 5
matrix(1:4, 2, 2) + matrix(1:4, 2, 2)
array(1:8, c(2, 2, 2)) + array(1:8, c(2, 2, 2))

## normal errors
Sys.Date() + Sys.Date()

## paste together characters
"this" + "that"

## or characters and numbers
"this" + 2000
2000 + "that"

## ignores missing
c(NA, "this") + c("that", NA) + c("other", "other")

## add space to behave like paste
"this" + " that"
"this " + "that"

## string together any number of character vectors
"this " + "that " + "and the other"
"ABC_" + letters + "_XYZ"
letters + letters
"T" + "his " + "that " + "and " + "the other."


}
