% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmsg.R
\name{pmsg}
\alias{pmsg}
\title{Paste collapse input and print as message}
\usage{
pmsg(..., print = TRUE)
}
\arguments{
\item{...}{Strings to be paste0 collapse = ""}

\item{print}{Logical indicating whether to print the message (default) or return an unevaluated expression}
}
\value{
Either invisibly returns text of message or unevaluated expression
}
\description{
Paste collapse input and print as message
}
