% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\name{print_as_col}
\alias{print_as_col}
\title{Print data frame as a single column}
\usage{
print_as_col(x)
}
\arguments{
\item{x}{Input data frame}
}
\value{
Prints first row and variable names
}
\description{
Prints first row of data frame
}
