% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-messages.R
\name{print_complete}
\alias{print_complete}
\title{Print message about completing a task}
\usage{
print_complete(...)
}
\arguments{
\item{...}{Strings collapsed (with no additional space added) into black
message prefixed with a heavy-check emoji (color and emoji print may
appear differently depending on your system/UI configuration)}
}
\value{
A printed message
}
\description{
Prints a check-mark bulleted message presumably about task completion
}
