% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-messages.R
\name{print_start}
\alias{print_start}
\title{Print message about starting a task}
\usage{
print_start(...)
}
\arguments{
\item{...}{Strings collapsed (with no additional space added) into gray
message prefixed with a next-arrow emoji (color and emoji print may
appear differently depending on your system/UI configuration)}
}
\value{
A printed message
}
\description{
Prints a next-arrow bulleted message presumably about task completion
}
