% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tbl.R
\name{print_tbl}
\alias{print_tbl}
\title{print_tbl}
\usage{
print_tbl(x, ..., title = TRUE)
}
\arguments{
\item{x}{Tibble}

\item{...}{Passed to print}

\item{title}{Name of data frame, defaults to "A tibble..."}
}
\value{
Print
}
\description{
Custom tweaks to tibble printing.
}
