% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psub.R
\name{psub}
\alias{psub}
\title{Paste sub}
\usage{
psub(x, ...)
}
\arguments{
\item{x}{Input string}

\item{...}{Named strings with names being the values to replace and the strings
being the desired new value.}
}
\description{
Glue-like sub pasting of strings
}
