% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{read_json_text}
\alias{read_json_text}
\title{read_json}
\usage{
read_json_text(x)
}
\arguments{
\item{x}{Character vector with json strings}
}
\value{
Parsed output from jsonlite fromJSON.
}
\description{
read_json
}
