% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-save-rds.R
\name{read_RDS}
\alias{read_RDS}
\title{Read RDS}
\usage{
read_RDS(path)
}
\arguments{
\item{path}{Name of .rds file}
}
\value{
A data object
}
\description{
Read serialized R data file
}
\seealso{
Other readsave: \code{\link{save_RDS}}
}
\concept{readsave}
