% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{readline_}
\alias{readline_}
\title{readline_}
\usage{
readline_(...)
}
\arguments{
\item{...}{Character string or vector to be used as prompt during
interactive R session. Ultimately, this function only sends a
single string to the user, but it will accept a vector if you're
picky about not creating strings of a certain width.}
}
\value{
Input entered during interactive session without extra quotes.
}
\description{
Worry free way to read lines from interactive sessions.
}
