% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{readlines}
\alias{readlines}
\title{readlines}
\usage{
readlines(x, ...)
}
\arguments{
\item{x}{Input}

\item{...}{Other args passed to \code{readLines}.}
}
\value{
Output
}
\description{
Read lines of file
}
\details{
Simple wrapper around \link[base]{readLines} that automates opening and
  closing of connection file.
}
