% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readr_cols.R
\name{readr_cols}
\alias{readr_cols}
\title{Easy readr column specification}
\usage{
readr_cols(...)
}
\arguments{
\item{...}{All arguments are passed to \code{\link[readr]{readr::cols}} with
any character values first treated and processed as shortcuts for equivalent
 \pkg{readr} functions–i.e., \code{readr::col_*}, \code{readr::parse_*}.
 See details for a list of available shortcuts.}
}
\value{
Column specification
}
\description{
Shortcut-loaded convenience wrapper around \code{\link[readr]{cols}}.
}
\details{
The following list of shortcuts can be supplied as character strings
  in place of \pkg{readr} \code{col_*} or \code{parse_*} functions:

\itemize{
  \item l, lgl, logical
  \item i, int, integer
  \item d, dbl, double
  \item c, chr, character
  \item D, date
  \item T, dt, dttm, datetime, POSIXct
  \item f, fct, factor
  \item n, num, number
  \item ?, gss, guess
  \item _, -, skip
}
}
\examples{

## readr package must be installed
if (requireNamespace("readr"), quietly = TRUE) {
  ## create column specification
  readr_cols(
    .default = "?",
    id = "chr",
    is_good = "lgl",
    day = "date",
    timestamp = readr::col_datetime(format = "\%m-\%y-\%d \%H:\%M:\%S"),
    count = "integer",
    misc = "_"
  )
}
}
