% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmatches_.R
\name{regmatches_}
\alias{regmatches_}
\alias{regmatches_first}
\title{extract reg expr matches}
\usage{
regmatches_(x, pat, drop = FALSE, ...)

regmatches_first(x, pat, drop = FALSE, ...)
}
\arguments{
\item{x}{Text data.}

\item{pat}{Reg ex pattern}

\item{drop}{Logical indicating whether to drop empty matches. Defaults to FALSE.}

\item{...}{Other args (like ignore.case) passed to gregexpr}
}
\value{
Matching expression from text.
}
\description{
A wrapper around the base function combo of gregexpr and regmatches
}
