% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitrename.R
\name{rename_git_repo}
\alias{rename_git_repo}
\title{Rename git repo code}
\usage{
rename_git_repo(new_url = NULL)
}
\arguments{
\item{new_url}{Optional, URL pointing to correct (renamed) repo. If NULL
(default), "new_url" is printed in brackets.}
}
\value{
text of git command to rename git repo
}
\description{
Command line (bash) syntax for renaming a git repo
}
