% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsstore.R
\name{rm_.DS_Store}
\alias{rm_.DS_Store}
\title{Remove pesky .DS_Store files}
\usage{
rm_.DS_Store()
}
\description{
Recursively removes all .DS_Store files in working directory.
}
\examples{

\dontrun{
rm_.DS_Store()
}

}
