% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_lines.R
\name{rm_amp}
\alias{rm_amp}
\title{rm_amp}
\usage{
rm_amp(x, y = "&")
}
\arguments{
\item{x}{Input character vector}

\item{y}{Replacement text}
}
\value{
Output character vector
}
\description{
Remove ampersands as they show up in tweet data
}
