% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_lines.R
\name{rm_linebreaks}
\alias{rm_linebreaks}
\title{rm_linebreaks}
\usage{
rm_linebreaks(x, y = " ")
}
\arguments{
\item{x}{Input character vector}

\item{y}{Replacement text}
}
\value{
Output character vector
}
\description{
Remove linebreaks
}
