% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.R
\name{rm_links}
\alias{rm_links}
\title{rm_links}
\usage{
rm_links(x)
}
\arguments{
\item{x}{Character vector of Twitter statuses.}
}
\value{
Character vector of statuses without URLs.
}
\description{
Removes URL links included in tweets.
}
