% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_lines.R, R/rm.R
\name{rm_retweets}
\alias{rm_retweets}
\alias{rm_retweets}
\title{rm_retweets}
\usage{
rm_retweets(x, ...)

rm_retweets(x, ...)
}
\arguments{
\item{x}{Input, should be either data frame (or list) or character.
If list/data frame, the text of the tweet is assumed to be named
"text" while the logical is_retweet is assuemd to be named "is_retweet"}

\item{x}{Input, should be either data frame (or list) or character.
If list/data frame, the text of the tweet is assumed to be named
"text" while the logical is_retweet is assuemd to be named "is_retweet"}
}
\value{
Output

Output
}
\description{
Remove retweets from data.

Remove retweets from data.
}
