% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.R
\name{rm_stopwords}
\alias{rm_stopwords}
\title{rm_stopwords}
\usage{
rm_stopwords(x, stopwords)
}
\arguments{
\item{x}{Vector of text.}

\item{stopwords}{Optional, stop words to be removed from text. Defaults to
SMART stop words provided by tidytext package.}
}
\value{
Character vector with stopwords removed
}
\description{
Returns statuses with stop words removed
}
