% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rproj.R
\name{rproj}
\alias{rproj}
\title{rproj}
\usage{
rproj(name = NULL, pkg = TRUE, ...)
}
\arguments{
\item{name}{Name of project. Defaults to basename of working directory.}

\item{pkg}{Logical indicating whether the project is an R package. Defaults
to true.}

\item{...}{Other args passed to rproj settings.}
}
\value{
Saved .Rproj file in current working directory.
}
\description{
Creates new rProf file.
}
