% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_normal}
\alias{scale_normal}
\title{Normalize scale(s)}
\usage{
scale_normal(x)
}
\arguments{
\item{x}{Data}
}
\value{
Numeric objects transformed via standard normal distribution.
}
\description{
Transform numeric objects into z-scores
}
