% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_standard}
\alias{scale_standard}
\title{Standardize scale(s)}
\usage{
scale_standard(x)
}
\arguments{
\item{x}{Data}
}
\value{
Numeric objects transformed onto 0-1 scale
}
\description{
Transform numeric objects into values on 0-1 scale
}
