% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_links.R
\name{scrape_links}
\alias{scrape_links}
\title{scrape URL links from a webpage}
\usage{
scrape_links(rurl, qs = TRUE, pat = ".")
}
\arguments{
\item{rurl}{URL to desired page. Should include scheme (http/https)}

\item{qs}{Logical indicating whether to return query strings (part of URL starting with ?)}

\item{pat}{Pattern from which only matching results should be returned. Defaults to ".",
which matches everything.}
}
\value{
Character vector of URLs.
}
\description{
Scrapes URL links from HTML source code
}
