% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_function.R
\name{search_function}
\alias{search_function}
\title{search functions}
\usage{
search_function(pat, fun)
}
\arguments{
\item{pat}{Pattern to match}

\item{fun}{Function to inspect}
}
\description{
Like search_files but for functions
}
