% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_files.R
\name{search_these_files}
\alias{search_these_files}
\title{Search these files}
\usage{
search_these_files(x, f)
}
\arguments{
\item{x}{Regex pat}

\item{f}{Vector of files}
}
\value{
Prints matches
}
\description{
Look for text in a group of files
}
