% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class.R
\name{set_class}
\alias{set_class}
\alias{add_class}
\title{set class}
\usage{
set_class(x, value)

add_class(x, value)
}
\arguments{
\item{x}{Object to assign new class to.}

\item{value}{Class value to assign to x}
}
\value{
Object x as class value.
}
\description{
Set class with a parenthetical function.

Add class with a parenthetical function.
}
