% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{set_description_params}
\alias{set_description_params}
\title{set_description_params}
\usage{
set_description_params(pkg = ".", title = NULL, description = NULL)
}
\arguments{
\item{pkg}{Name of package (basename of wd)}

\item{title}{Title of package.}

\item{description}{Description of package.}
}
\value{
Stores description info for devtools.
}
\description{
Sets desc options for devtools
}
