% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{set_names}
\alias{set_names}
\title{Set names}
\usage{
set_names(x, nms)
}
\arguments{
\item{x}{Data object}

\item{nms}{Names to assign object.}
}
\description{
Add names with a parenthetical function.
}
