% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{set_renv}
\alias{set_renv}
\title{set_renv}
\usage{
set_renv(...)
}
\arguments{
\item{...}{Named values where names correspond to the name of the
environment variables. Must include names for each value.}
}
\value{
Saves and reads into current session new environment variable(s).
}
\description{
Create and save new R environment variable
}
