% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shhh.R
\name{shhh}
\alias{shhh}
\title{execute expression quietly}
\usage{
shhh(expr)
}
\arguments{
\item{expr}{Expression to be evaluated without additional printing.}
}
\value{
Output from evaluated expression.
}
\description{
execute expression quietly
}
