% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{should_be_post}
\alias{should_be_post}
\title{should_be_post}
\usage{
should_be_post(.query, .nchar = 60)
}
\arguments{
\item{x}{api params}
}
\value{
logical indicating whether the query exceeds the specified
cutoff point
}
\description{
should_be_post
}

