% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_connections.R
\name{show_connections}
\alias{show_connections}
\title{show connections}
\usage{
show_connections()
}
\value{
Prints and invisibly returns data frame.
}
\description{
Displays active connections as a tidy tibble
}
