% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_ids.R
\name{sn2id}
\alias{sn2id}
\title{sn2id}
\usage{
sn2id(screen_name, mc.cores = 1L)
}
\arguments{
\item{screen_name}{Twitter handle}

\item{mc.cores}{Number of cores. If greater than 1, then mcMap is used.}
}
\value{
response Twitter account user id
}
\description{
sn2id
}
