% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliturl.R
\name{split_url}
\alias{split_url}
\title{Split URL}
\usage{
split_url(x)
}
\arguments{
\item{x}{Input character vector}
}
\value{
If input is one string, then a character vector. If multiple URLs,
then a list of character vectors
}
\description{
Splits URL(s) such that each URL is represented with a character vector split
at all single forward slashes and ampersand (forward slashes and ampersands
are included in the return object)
}
