% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_section.R
\name{square_section}
\alias{square_section}
\title{square_section}
\usage{
square_section()
}
\description{
Load clipboard with code chunk square. Paste to
  insert square into R script file at cursor location.
}

