% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{stripped_down_words}
\alias{stripped_down_words}
\title{stripped_down_words}
\usage{
stripped_down_words(x)
}
\arguments{
\item{x}{CHaracter vector of text to clean.}
}
\value{
Split into vectors of words for each character string.
}
\description{
Strip (clean) and tokenize text
}
\examples{
x <- "Some sentencE with AWK. formatting
and \\t a URL link to https://google.com   , maybe?"
stripped_down_words(x)
}
