% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrev.R
\name{substrev}
\alias{substrev}
\title{Sub string reverse}
\usage{
substrev(x, start, stop = 0)
}
\arguments{
\item{x}{Character vector}

\item{start}{Number of characters to include relative to the last
character position.}

\item{stop}{Specify the number of characters from the final
character to set as the last character position.}
}
\value{
Sub string with last i characters.
}
\description{
Returns portion of string starting from end of string (otherwise just like substr)
}
