% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{swp}
\alias{swp}
\title{stop words by probability}
\usage{
swp(p = 0.99999, lang = "en", drop = TRUE, data = NULL)
}
\arguments{
\item{p}{Probability of word in any given sample of tweets.}

\item{lang}{One or more lang abbreviations to subset with.}

\item{drop}{Logical indicating whether to drop non-word elements of
data frame (the default) or to return the subsetted [and ordered
by desc p] data set}

\item{data}{Data set of stopwords. Defaults to stopwordslangs from
rtweet package}
}
\value{
Either vector of words or data frame with word, p, lang.
}
\description{
Return Twitter specific stopwords by expected probability of a term
being unique to a given sample of tweets.
}
