% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{as_tbl}
\alias{as_tbl}
\alias{tbl_df}
\alias{data_frame}
\alias{as_tibble}
\title{tbl}
\usage{
as_tbl(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{as_tibble}}, \code{\link{tbl_df}},
or \code{\link{data_frame}}.}

\item{validate}{Logical indicating whether to validate tibble. Defaults to
FALSE because it's slightly faster.}
}
\value{
A tibble data frame
}
\description{
Tibble functions
}
\examples{
as_tbl(data.frame(x = rnorm(10), y = rnorm(10)))
tbl_df(mtcars)
data_frame(x = rnorm(10), y = rnorm(10)))
as_tibble(mtcars)
}
