% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mwk.R
\name{theme_chartist}
\alias{theme_chartist}
\title{theme_chartist}
\usage{
theme_chartist(base_size = 12, base_family = "Roboto Condensed")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Font family.}

\item{dark}{Primary forefront color.}

\item{light}{Primary background color.}

\item{gray}{Gray color.}
}
\value{
My ggplot2 theme (similar to theme_minimal/theme_bw)
}
\description{
A chartist inspired ggplot2 theme
}
