% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/this_in_that.R
\name{this_in_that}
\alias{this_in_that}
\title{Where is this in that?}
\usage{
this_in_that(this, that, value = NULL)
}
\arguments{
\item{this}{Values to look up in that}

\item{that}{Value positions matched to this}

\item{value}{Optional, values to be returned rather than the default, which
returns positions (integers)}
}
\description{
Looks up (matches) the position of \code{this} in \code{that} (table)
}
