% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{tibble}
\alias{tibble}
\alias{as_tbl}
\alias{tbl_df}
\title{tibble}
\usage{
as_tbl(x, ...)

tbl_df(...)
}
\description{
See \code{tibble::\link[tibble]{as_tibble}} for details.

See \code{tibble::\link[tibble]{as_tibble}} for details.
}
\keyword{internal}
