% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezones.R
\name{time2utc}
\alias{time2utc}
\title{convert date-time time zone to UTC}
\usage{
time2utc(x)
}
\arguments{
\item{x}{data}
}
\value{
Object with date-times in UTC timezone
}
\description{
convert date-time time zone to UTC
}
\examples{
time2utc(Sys.time())
}
