% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titleupper.R
\name{titleupper}
\alias{titleupper}
\title{titleupper}
\usage{
titleupper(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector with non-article/preposition (unless at beginning of sentence of following colon) first letters capitalized
}
\description{
Capitalize characters to title case
}
\examples{
titleupper("this is a title about something good: a witty zinger")
}
