% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.R
\name{tl}
\alias{tl}
\title{Tail data}
\usage{
tl(x, i = 10)
}
\arguments{
\item{x}{Input}

\item{i}{Number of elements to return from last, defaults to 10.}
}
\value{
Data with last i elements.
}
\description{
Returns last i elements of data object.
}
