% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_used.R
\name{token_used}
\alias{token_used}
\title{Calculate and sort by API use}
\usage{
token_used(x)
}
\arguments{
\item{x}{Parsed token data}
}
\value{
Arrange data frame with calculated used column.
}
\description{
Counts number of requests used for each API (observation) and
arranges in descending order
}
