% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_ws.R
\name{trim_ws}
\alias{trim_ws}
\title{trim_ws}
\usage{
trim_ws(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector without extra spaces
}
\description{
Returns character vector without extra spaces and trimmed of white space.
}
