% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetshot.R
\name{tweet_url}
\alias{tweet_url}
\title{tweet url}
\usage{
tweet_url(x)
}
\arguments{
\item{x}{Data with status ID and screen name variables.}
}
\value{
URLs to statuses.
}
\description{
Returns URL to twitter status
}
