% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.functions.R
\name{TWIT}
\alias{TWIT}
\title{TWIT}
\usage{
TWIT(query, parameters, token)
}
\arguments{
\item{query}{Twitter API request type string. e.g, \code{"friends/ids"} calls Twitter API to return information about a user's friend network (i.e., accounts followed by a user).}

\item{parameters}{Additional parameters passed along to API call
See \url{https://dev.twitter.com/overview/documentation} for more information on using Twitter's API.}

\item{token}{An OAuth token (1.0 or 2.0)}

\item{type}{If not missing, which returns entire rate limit request object, type returns specific values; e.g., \code{type = "lookup"} returns remaining limit for user lookup requests; \code{type = "followers"} returns remaining limit for follower id requests; \code{type = "friends"} returns remaining limit for friend id requests.}
}
\value{
json response object as nested list
}
\description{
TWIT
}

