% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezones.R
\name{utc_time}
\alias{utc_time}
\alias{cst_time}
\alias{est_time}
\alias{mst_time}
\alias{pst_time}
\title{UTC time zone}
\usage{
utc_time(x = Sys.time())

cst_time(x = Sys.time())

est_time(x = Sys.time())

mst_time(x = Sys.time())

pst_time(x = Sys.time())
}
\arguments{
\item{x}{}
}
\value{
Date time in UTC time zone
}
\description{
UTC time zone

CST time zone

EST time zone

MST time zone

PST time zone
}
\examples{
utc_time(Sys.time())
ets_time(Sys.time())
cst_time(Sys.time())
mst_time(Sys.time())
pst_time(Sys.time())
}
