% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip.R
\name{un_zip}
\alias{un_zip}
\title{Unzip files into directory}
\usage{
un_zip(path)
}
\arguments{
\item{path}{Name of zipfile. Must end in ".zip"}
}
\value{
Creates a directory in the same folder with the same name (minus the zip part)
}
\description{
Unzip archive into similarly named directory
}
