% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_number.R
\name{update_citation}
\alias{update_citation}
\title{update_citation}
\usage{
update_citation(version, path = ".")
}
\arguments{
\item{version}{New version number.}

\item{path}{Path to pkg, defaults to "."}
}
\description{
Updates package citation file
}
