% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tweet_text.R
\name{update_tweet_text}
\alias{update_tweet_text}
\title{update_tweet_text}
\usage{
update_tweet_text(x)
}
\arguments{
\item{x}{Data frame of tweets data.}
}
\value{
Data frame with updated text column
}
\description{
Update text variable to include retweet/quote text as relevant
}
