% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_number.R
\name{update_version_number}
\alias{update_version_number}
\title{update_version_number}
\usage{
update_version_number(update_type = "patch", pkg = ".")
}
\arguments{
\item{update_type}{Patch, minor, or major update.}

\item{path}{Path to base directory of given R package.}
}
\value{
Updates and saves DESCRIPTION file.
}
\description{
Updates version number in R package documentation
}
