% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_number.R
\name{version_number}
\alias{version_number}
\title{version_number}
\usage{
version_number(major, minor, patch)
}
\arguments{
\item{major}{Increment major, e.g. 1.0.0, for a major release. This
is best reserved for changes that are not backward compatible and
that are likely to affect many users. Going from 0.b.c to 1.0.0
typically indicates that your package is feature complete with a
stable API.}

\item{minor}{Increment minor, e.g. 0.9.0, for a minor release. A
minor release can include bug fixes, new features and changes in
backward compatibility. This is the most common type of
release. It’s perfectly fine to have so many minor releases that
you need to use two (or even three!) digits, e.g. 1.17.0.}

\item{major}{Increment patch, e.g. 0.8.2 for a patch: you’ve fixed
bugs without adding any significant new features. I’ll often do a
patch release if, after release, I discover a show-stopping bug
that needs to be fixed ASAP. Most releases will have a patch
number of 0.}
}
\description{
Class for version numbers.
}
