% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{wave_path_name}
\alias{wave_path_name}
\title{wave_path_name}
\usage{
wave_path_name(data_folder = "/Users/mwk/r/tfse/study/data/")
}
\value{
Printed output via cat() function for copy/paste of long strings
}
\description{
wave_path_name
}

