% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.friends.R
\name{which_ids}
\alias{which_ids}
\title{which_ids}
\usage{
which_ids(n)
}
\arguments{
\item{n}{starting number for users}
}
\value{
integers used to identify 15 users
}
\description{
Returns integer values. Used for get_friends function.
}

