% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.x.R
\name{with.x}
\alias{with.x}
\title{do something with .x}
\usage{
\method{with}{x}(.x, expr)
}
\arguments{
\item{.x}{Data}

\item{expr}{Expression (the thing to do)}
}
\value{
Evaluates expr with .x as environment
}
\description{
do something with .x
}
