% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliturl.R
\name{wrap_url}
\alias{wrap_url}
\title{Wrap URL}
\usage{
wrap_url(x, width = NULL)
}
\arguments{
\item{x}{Input URL}

\item{width}{Width defaulting to 80 (or current width if smaller)}
}
\value{
Text copied to clipboard and printed
}
\description{
Wraps long URL with the %P% operator
}
