% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-save-rds.R
\name{write_RDS}
\alias{write_RDS}
\title{Read RDS}
\usage{
write_RDS(x, path, compress = FALSE)
}
\arguments{
\item{x}{Data object to be saved}

\item{path}{Name of the file to be saved}

\item{compress}{Logical indicating whether to compress data–it will take up
less space but it will be slower. Defaults to FALSE.}
}
\value{
Invisible data object
}
\description{
Read serialized R data file
}
\seealso{
Other readsave: \code{\link{read_RDS}}
}
\concept{readsave}
