% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{writelines}
\alias{writelines}
\title{writelines}
\usage{
writelines(x, ...)
}
\arguments{
\item{x}{Character, file name to save as.}

\item{...}{Other args passed to \code{writeLines}.}
}
\value{
Saves/overwrites to file.
}
\description{
Writes lines to file
}
