% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{add_to_git}
\alias{add_to_git}
\title{add_to_git}
\usage{
add_to_git(m = "misc", pkg = ".", pull = TRUE, interactive = TRUE)
}
\arguments{
\item{m}{Commit message}

\item{pkg}{Repo name. Defaults to working directory file name.}

\item{pull}{Logical indicating whether to pull the repo prior to
pushing}

\item{interactive}{Logical indicating whether to ask prior to
committing to Github repo.}
}
\value{
Sends local repo to Github.
}
\description{
Update github repository
}
