% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_section.R
\name{boxcode}
\alias{boxcode}
\title{boxcode}
\usage{
boxcode(label = "")
}
\arguments{
\item{label}{Name of section/block}
}
\value{
Text for code box saved into clipboard. Paste to use at cursor.
}
\description{
Load clipboard with code chunk square. Paste to insert square into R script 
  file at cursor location.
}
