% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.functions.R
\name{get_friends}
\alias{get_friends}
\title{get_friends}
\usage{
get_friends(user, token, page = "-1")
}
\arguments{
\item{token}{OAuth token (1.0 or 2.0)}

\item{page}{Default \code{page = -1} specifies first page of json results. Other pages specified via cursor values supplied by Twitter API response object.}

\item{screen_name}{Screen name of target user.}
}
\value{
friends User ids for everyone a user follows.
}
\description{
get_friends
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more information on using Twitter's API.
}

