% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.functions.R
\name{get_friends_max}
\alias{get_friends_max}
\title{get_friends_max}
\usage{
get_friends_max(ids, tokens, group, start)
}
\arguments{
\item{tokens}{OAuth tokens (1.0 or 2.0)}

\item{followers}{Data frame with column name "screen_name"}
}
\value{
friends List of user ids each user follows.
}
\description{
get_friends_max
}

