% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.friends.R
\name{get_friends_max}
\alias{get_friends_max}
\title{get_friends_max}
\usage{
get_friends_max(user_ids, tokens, start = 1, stringify = TRUE,
  verbose = TRUE)
}
\arguments{
\item{user_ids}{Data frame with column name "screen_name"}

\item{tokens}{OAuth tokens (1.0 or 2.0)}

\item{start}{Starting value (nth user id)}

\item{stringify}{logical, indicating whether to return user ids as strings}

\item{verbose}{default behavior \code{verbose = TRUE} prints asterisk for every 15 (or max of one token) user networks collected. Set \code{verbose = FALSE} to run function silently.
(some ids are too long to be read as numeric). Defaults to \code{TRUE}.}
}
\value{
friends List of user ids each user follows.
}
\description{
get_friends_max
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more
  information on using Twitter's API.
}

