% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.friends.R, R/get.statuses.R
\name{get_friendships}
\alias{get_friendships}
\title{get_friendships}
\usage{
get_friendships(user, since_id = NULL, count = NULL, max_id = NULL,
  trim_user = NULL, exclude_replies = NULL, contributer_details = NULL,
  include_rts = NULL, token)

get_friendships(user, since_id = NULL, count = NULL, max_id = NULL,
  trim_user = NULL, exclude_replies = NULL, contributer_details = NULL,
  include_rts = NULL, token)
}
\arguments{
\item{user}{The screen_name or ID of the user for whom to return results for.}

\item{since_id}{Returns results with an ID greater than (that is, more recent than) the specified ID. There are limits to the number of Tweets which can be accessed through the API. If the limit of Tweets has occured since the since_id, the since_id will be forced to the oldest ID available.}

\item{count}{Specifies the number of tweets to try and retrieve, up to a maximum of 200 per distinct request. The value of count is best thought of as a limit to the number of tweets to return because suspended or deleted content is removed after the count has been applied. We include retweets in the count, even if include_rts is not supplied. It is recommended you always send include_rts=1 when using this API method.}

\item{max_id}{Returns results with an ID less than (that is, older than) or equal to the specified ID.}

\item{trim_user}{When set to either true, t or 1, each tweet returned in a timeline will include a user object including only the status authors numerical ID. Omit this parameter to receive the complete user object.}

\item{exclude_replies}{This parameter will prevent replies from appearing in the returned timeline. Using exclude_replies with the count parameter will mean you will receive up-to count tweets — this is because the count parameter retrieves that many tweets before filtering out retweets and replies. This parameter is only supported for JSON and XML responses.}

\item{include_rts}{When set to false, the timeline will strip any native retweets (though they will still count toward both the maximal length of the timeline and the slice selected by the count parameter). Note: If you’re using the trim_user parameter in conjunction with include_rts, the retweets will still contain a full user object.}

\item{token}{OAuth token (1.0 or 2.0)}

\item{source}{Screen name or user id of source user.}

\item{target}{Screen name or user id of target user.}

\item{contributor_details}{This parameter enhances the contributors element of the status response to include the screen_name of the contributor. By default only the user_id of the contributor is included.}

\item{token}{OAuth token (1.0 or 2.0)}
}
\value{
json object (nested list)

json object (nested list)
}
\description{
Returns detailed information about the relationship between two arbitrary users.

Returns a collection of the most recent Tweets posted by the user indicated by the screen_name or user_id parameters.
User timelines belonging to protected users may only be requested when the authenticated user either “owns” the timeline or is an approved follower of the owner.
The timeline returned is the equivalent of the one seen when you view a user’s profile on twitter.com.
This method can only return up to 3,200 of a user’s most recent Tweets. Native retweets of other statuses by the user is included in this total, regardless of whether include_rts is set to false when requesting this resource.
}
\seealso{
See \url{https://api.twitter.com/1.1/friendships/show.json}.

See \url{https://api.twitter.com/1.1/statuses/user_timeline.json}.
}

