% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.lookup.R
\name{get_lookup}
\alias{get_lookup}
\title{get_lookup}
\usage{
get_lookup(users, token, df = TRUE)
}
\arguments{
\item{users}{Screen name or user id of target users.}

\item{token}{OAuth tokens (1.0 or 2.0)}

\item{df}{logical, indicating whether to format response as data frame}
}
\value{
response object
}
\description{
get_lookup
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more
  information on using Twitter's API.
}

