% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.lookup.R
\name{get_lookup_max}
\alias{get_lookup_max}
\title{get_lookup_max}
\usage{
get_lookup_max(ids, tokens, start = 1)
}
\arguments{
\item{ids}{User id or screen name of target user.}

\item{tokens}{OAuth tokens (1.0 or 2.0)}

\item{start}{First (nth) id}
}
\value{
response object
}
\description{
get_lookup_max
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more
  information on using Twitter's API.
}

