% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id2sn}
\alias{id2sn}
\title{id2sn}
\usage{
id2sn(user_id, mc.cores = 1L)
}
\arguments{
\item{user_id}{Vector of user_id(s)}
}
\value{
Named vector of screen names.
}
\description{
Converts user ids to screen names.
}
