% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\name{n_uq}
\alias{n_uq}
\title{Number of unique elements}
\usage{
n_uq(x)
}
\arguments{
\item{x}{Input data}
}
\value{
Integer number of unique elements
}
\description{
Estimates number of unique elements in a data object
}
