% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-purrr.R
\name{pluck}
\alias{pluck}
\title{Pluck out a single an element from a vector or environment}
\usage{
pluck(.x, .f)
}
\arguments{
\item{.x}{A vector or environment}

\item{.f}{Accessors for indexing into the object. Can be an integer position, a string name, or an accessor function.}
}
\description{
Pluck out a single an element from a vector or environment
}
