% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rproj.R
\name{rproj_pkg}
\alias{rproj_pkg}
\alias{rproj_site}
\title{rproj}
\usage{
rproj_pkg(path = NULL)

rproj_site(path = NULL)
}
\arguments{
\item{path}{Project name (path).}
}
\value{
Saved .Rproj file in path.
}
\description{
Creates new .Rproj file.
}
