% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_files.R
\name{search_files}
\alias{search_files}
\title{search_files}
\usage{
search_files(x, path = ".", recursive = TRUE, all.files = FALSE)
}
\arguments{
\item{x}{Pattern.}

\item{path}{Path on which to restrict search. Defaults to current working
directory.}

\item{recursive}{logical}

\item{all.files}{default false excludes dot files}
}
\value{
Output from terminal - file name, line number, and preview of
  matching text
}
\description{
Returns matching files and line numbers of given string pattern.
}
