% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{set_renv}
\alias{set_renv}
\title{Set R environment variable}
\usage{
set_renv(...)
}
\arguments{
\item{...}{Named environment variables/values}
}
\value{
Appends environment variable entry to ~/.Renviron file
}
\description{
Sets R environment variable and adds it to user's home .Renviron file
}
