% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sn2id}
\alias{sn2id}
\title{sn2id}
\usage{
sn2id(screen_name)
}
\arguments{
\item{screen_name}{Twitter handle}
}
\value{
response Twitter account user id
}
\description{
sn2id
}
\details{
rvest
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more information on using Twitter's API.
}

