% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{stopwords}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 571 rows and 2 columns.}
\usage{
stopwords
}
\description{
Vector of stopwords taken from SMART lexicon in tidytext
}
\keyword{datasets}
