% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{tabsort}
\alias{tabsort}
\title{tabsort}
\usage{
tabsort(x, V1 = NULL, percent = TRUE)
}
\arguments{
\item{x}{Character vector}

\item{V1}{Optional, name of term variable. Defaults to "term".}

\item{percent}{Logical indicating whether to include a percent of total
column.}
}
\value{
Frequency tbl
}
\description{
Returns a sorted (descending) frequence tbl
}
