% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabsort.R
\name{tabsort}
\alias{tabsort}
\alias{ntbl}
\title{tabsort}
\usage{
tabsort(data, ..., prop = TRUE, na_omit = TRUE, sort = TRUE)

ntbl(data, ...)
}
\arguments{
\item{data}{Data}

\item{prop}{Logical indicating whether to include a proportion of total
obs column.}

\item{na_omit}{Logical indicating whether to exclude missing. If all
responses are missing, a missing value is used as the single category.}
}
\value{
Frequency tbl
}
\description{
Returns a sorted (descending) frequence tbl
}
