% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{tbl}
\alias{tbl}
\alias{as_tbl}
\alias{tbl_frame}
\alias{as_tbl_smart}
\alias{tblframe}
\title{tbl}
\usage{
tbl(x, row.names = FALSE)

as_tbl(x, row.names = FALSE)

tbl_frame(...)

as_tbl_smart(x)

tblframe(x)
}
\arguments{
\item{x}{Data}

\item{row.names}{Logical indicating whether to convert non-null row names
into the first column.}
}
\description{
See \code{tibble::\link[tibble]{as_tibble}} for details.

See \code{tibble::\link[tibble]{as_tibble}} for details.
}
