% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfse.R
\docType{package}
\name{tfse}
\alias{tfse}
\alias{tfse-package}
\title{tfse: Twitter Follows and Selective Exposure}
\description{
Useful user functions and imported functions.
}
\author{
Michael W Kearney
}
\seealso{
\code{httr}
}
\keyword{package}
\keyword{tfse-package}

