% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfse.R
\docType{package}
\name{tfse}
\alias{tfse}
\alias{tfse-package}
\title{tfse: Twitter Follows and Selective Exposure}
\description{
An implementation of http calls and data wrangling
functions designed to interact with Twitter API's. GET and
POST requests are designed to return data from Twitter's
search and streaming API's. Returned [json] objects are
automatically parsed into data_frames or lists. Specific
consideration is given within the functions to use of more
than one OAuth personal access token.
}

