% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfse.R
\docType{package}
\name{tfse}
\alias{tfse}
\alias{tfse-package}
\title{tfse: Various Useful Functions}
\description{
Collection of useful functions.
}
\author{
Michael W. Kearney
}
\keyword{package}
\keyword{tfse-package}
