% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mwk.R
\name{theme_mwk}
\alias{theme_mwk}
\title{theme_mwk}
\usage{
theme_mwk(base_size = 14, base_family = "Roboto", dark = "#24292e",
  light = "#ffffff", gray = "#f0f0f0", ...)
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Font family.}

\item{dark}{Primary forefront color.}

\item{light}{Primary background color.}

\item{gray}{Gray color.}
}
\value{
My ggplot2 theme (similar to theme_minimal/theme_bw)
}
\description{
My ggplot2 theme
}
