% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mwk.R
\name{theme_mwk}
\alias{theme_mwk}
\alias{theme_mwk_caption_text}
\title{theme_mwk}
\usage{
theme_mwk(base_size = 12, base_family = "Roboto Condensed",
  dark = "#24292e", light = "#ffffff", gray = "#ededed",
  caption = TRUE)

theme_mwk_caption_text()
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Font family.}

\item{dark}{Primary forefront color.}

\item{light}{Primary background color.}

\item{gray}{Gray color.}

\item{caption}{Text to use as source caption.}
}
\value{
My ggplot2 theme (similar to theme_minimal/theme_bw)
}
\description{
My ggplot2 theme
}
