% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.tweets.R
\name{top_tweet_words}
\alias{top_tweet_words}
\title{top_tweet_words}
\usage{
top_tweet_words(tweets_text, min = 3)
}
\arguments{
\item{tweets_text}{character vector of tweets text}

\item{min}{minimum number of ocurrences to include in returned object}
}
\value{
list object with top mentions and top words
}
\description{
top_tweet_words
}

