% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.functions.R
\name{TWIT}
\alias{TWIT}
\title{TWIT}
\usage{
TWIT(query, parameters = NULL, token, parse = TRUE, version = "1.1")
}
\arguments{
\item{query}{Twitter API request type string. e.g, \code{"friends/ids"} calls
Twitter API to return information about a user's friend network (i.e.,
accounts followed by a user).}

\item{parameters}{Additional parameters passed along to API call}

\item{token}{An OAuth token (1.0 or 2.0)}
}
\value{
json response object as nested list
}
\description{
TWIT
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation} for more
  information on using Twitter's API.
}

