% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_function.R
\name{write_function}
\alias{write_function}
\title{Write function to file}
\usage{
write_function(...)
}
\arguments{
\item{...}{Functions to write to file. It's best to name these.}
}
\value{
Writes to temporary file and opens that file.
}
\description{
Write the function code to a file and open the file
}
