% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{writelines}
\alias{writelines}
\title{writelines}
\usage{
writelines(x, file, ...)
}
\arguments{
\item{x}{Text to output.}

\item{file}{File name}

\item{...}{Other args passed to \code{writeLines}.}
}
\value{
Saves/overwrites to file.
}
\description{
Writes lines to file
}
